test(auth): merge invalidation tests into hash cache test file, remove duplicate
